stardust-backend/requirements.txt
asgiref==3.9.1
dj-rest-auth==7.0.1
Django==4.2.23
django-allauth==65.10.0
django-axes==8.0.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-otp==1.6.1
django-ratelimit==4.1.0
django-rest-passwordreset==1.5.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.1
pillow==11.3.0
pkg_resources==0.0.0
psycopg2-binary==2.9.10
PyJWT==2.10.1
sqlparse==0.5.3
typing_extensions==4.14.1
